feat(service): add EstatisticasService and calcularEstatisticasDeTransacoes method

+++ src/main/java/com/alan/transacao_api/controller/dto/EstatisticasResponseDTO.java
@@ -1,7 +1,7 @@
 package com.alan.transacao_api.controller.dto;
 
 public record EstatisticasResponseDTO(
-        Integer count,
+        Long count,
         Double sum,
         Double avg,
         Double min,
